Remove MvtFirstPassedData
